### 2020-02-17 #18 @kbroman
Ten Years Reproducibility Challenge: Identifying essential genes by mutagenesis
**Original article:** Lamichhane G, Zignol M, Blades NJ, Geiman DE, Dougherty A, Broman KW, Bishai WR (2003) A post-genomic method for predicting essential genes at subsaturation levels of mutagenesis: application to Mycobacterium tuberculosis. Proc Natl Acad Sci USA 100:7213-7218 [doi:10.1073/pnas.1231432100](https://doi.org/10.1073/pnas.1231432100)

**Scientific domain:**  Genetics
**Programming language:**  R
**Suggested editor:**  @karthik
