### 2020-02-17 2f7917e
Add zenodo DOI badge
