### 2020-02-20 #18 @rougier
@kbroman Thanks, we'll soon assign an editor.
@ReScience/editors Can someone handle this submission (finding an editor or editing yourself) ? Note that for the Ten Years Reproducibility Challenge, only one reviewer is needed.
