### 2020-02-23 eeef718
Re-run with R version 3.6.2
