+++ reproduction/Tabs/tab6.tex
@@ -1,5 +1,5 @@
-% latex table generated in R 3.6.1 by xtable 1.8-4 package
-% Fri Feb 14 19:18:38 2020
+% latex table generated in R 3.6.2 by xtable 1.8-4 package
+% Sun Feb 23 08:56:40 2020
 \begin{tabular}{ccccc}
   \hline
   & & & \multicolumn{2}{c}{Probability (\%)}\\ MT \# & Rv \# & Gene description & original & reproduction\\ \hline
