+++ reproduction/Tabs/tab2.tex
@@ -1,5 +1,5 @@
-% latex table generated in R 3.6.1 by xtable 1.8-4 package
-% Fri Feb 14 19:18:37 2020
+% latex table generated in R 3.6.2 by xtable 1.8-4 package
+% Sun Feb 23 08:56:40 2020
 \begin{tabular}{cccr@{--}lr@{--}l}
   \hline
   & \multicolumn{2}{c}{Estimate (\%)} & \multicolumn{4}{c}{95\% credible interval}\\ Rule & original & reproduction & \multicolumn{2}{c}{original} & \multicolumn{2}{c}{reproduction}\\ \hline
