+++ article/content.tex
@@ -130,7 +130,7 @@ reconstructed without much difficulty.
 \section{Code modifications}
 
 The original analysis was performed with R version 1.5.1 (2002-06-17);
-the reproduction used R version 3.6.1 (2019-07-05). The analysis
+the reproduction used R version 3.6.2 (2019-12-12). The analysis
 scripts could be run with only one small modification. The output of
 the R function {\tt table()}, which counts the values in a
 categorical variable, is now an object of class {\tt table}, where as
@@ -181,7 +181,7 @@ In this section, I present the results of my reproduction of the analyses
 in \citet{lamichhane2003}. I am not sure what hardware or operating
 system I used for the original analyses, but output files indicate I
 was using R version 1.5.1. For the reproduction, I used a System76
-Oryx Pro laptop running Pop!\_OS Linux 19.10 and R version 3.6.1.
+Oryx Pro laptop running Pop!\_OS Linux 19.10 and R version 3.6.2.
 
 Figure~\ref{fig:fig1a} displays the
 reproduction of Figure~1a from \citet{lamichhane2003}. The left panel
