### 2020-02-23 e9c172d
oops hadn't re-compiled reproduction.html
 - also needed to re-compile article.pdf
