+++ reproduction/Tabs/tab6.tex
@@ -1,5 +1,5 @@
 % latex table generated in R 3.6.2 by xtable 1.8-4 package
-% Sun Feb 23 08:56:40 2020
+% Sun Feb 23 09:00:36 2020
 \begin{tabular}{ccccc}
   \hline
   & & & \multicolumn{2}{c}{Probability (\%)}\\ MT \# & Rv \# & Gene description & original & reproduction\\ \hline
