+++ reproduction/Tabs/tab4.tex
@@ -1,5 +1,5 @@
 % latex table generated in R 3.6.2 by xtable 1.8-4 package
-% Sun Feb 23 08:56:40 2020
+% Sun Feb 23 09:00:36 2020
 \begin{tabular}{lccc@{ (}r@{--}lc@{ (}r@{--}l}
   \hline
   & \multicolumn{2}{c}{Probability enriched (\%)} & \multicolumn{6}{c}{Est. \% essential}\\ Functional group & original & reproduction & \multicolumn{3}{c}{original} & \multicolumn{3}{c}{reproduction}\\ \hline
