### 2020-04-10 3d60281
Fix typo: "where as" -> "whereas"
