### 2020-07-06 #18 @pcarbo
@karthik I just want to be clear what you are looking for. Are these the instructions from [here](https://rescience.github.io/edit) that are relevant to me?

> During the review, reviewers are free to interact with the authors to ask for clarification or change in any file that is part of the submission. The main criterion for acceptance is either: Replication of the original research with a clear statement by the authors explaining why they think they have replicated the paper (same figures, same graphics, same behavior, etc). Non-replication of the original research with a clear statement by the authors explaining why they believe that it cannot be replicated. Source code without any accompanying article is also a criterion for rejection since we’re not human compilers (well not all of us at least).
```
