### 2020-07-06 #18 @karthik
@pcarbo Some might not be relevant to this challenge as Karl is replicating his own work a decade later. So you can ignore these. Your goal is to run the code and see if Karl's reproducibility report (the PDF) matches your experience. You are still welcome to contact (tag) the author with questions and clarifications.

Here is an example review: https://github.com/ReScience/submissions/issues/32#issuecomment-633739558
