### 2020-07-15 #18 @kbroman
@pcarbo and @lwa19, thanks so much for your detailed comments, and for your efforts to reproduce my reproduction! I really appreciate the feedback.

I can see that my reproduction was complicated and cringe-inducing in places. I had in mind leaving the original scripts and data files in their place exactly as they were and using a single R Markdown document to modify and re-run things; that was probably not the best choice.
