### 2020-07-15 9c44c29
Fixed typo in package install code
