### 2020-07-15 #18 @karthik
Thanks so much @lwa19 and @pcarbo! Very grateful for your review.

@kbroman I opened a PR with some small fixes to your README.
