### 2020-07-15 #18 @karthik
Noting the original request here (for my own reference) https://github.com/ReScience/ten-years/issues/1#issuecomment-546353681
