### 2020-07-16 #18 @karthik
@pcarbo and @lwa19 Can you please share your ORCID for the paper metadata? Once I have it I'll send another pull request to Karl's repository.
