### 2020-07-16 f9a3ba6
Re-build manuscript following Karthik's additions
