+++ article/metadata.tex
@@ -5,12 +5,12 @@
 \def \codeSWH{}
 \def \dataURL{https://github.com/kbroman/Paper_ReScience2020/}
 \def \dataDOI{}
-\def \editorNAME{}
-\def \editorORCID{}
-\def \reviewerINAME{}
-\def \reviewerIORCID{}
-\def \reviewerIINAME{}
-\def \reviewerIIORCID{}
+\def \editorNAME{Karthik Ram}
+\def \editorORCID{0000-0002-0233-1757}
+\def \reviewerINAME{Leah Wang}
+\def \reviewerIORCID{0000-0001-5905-8061}
+\def \reviewerIINAME{Peter Carbonetto}
+\def \reviewerIIORCID{0000-0003-1144-6780}
 \def \dateRECEIVED{}
 \def \dateACCEPTED{}
 \def \datePUBLISHED{}
