### 2020-07-16 a78a70f
Clarify a few things in the R Markdown document describing reproduction
- Following @lwa19's suggestions in
  <https://lwa19.github.io/Paper_ReScience2020/reproduction/reproduction.html>
