+++ reproduction/reproduction.Rmd
@@ -161,7 +161,10 @@ if(!file.exists(new_file)) {
 ## Run perl script
 
 I will first run the Perl script `findTA.pl` which uses data from
-`Rawdata/` and creates a set of files in `Data/`.
+`Rawdata/` and creates a set of files in `Data/`. More specifically,
+it uses the genomic sequence for _M. tuberculosis_ and the genes'
+start and end positions to identify the locations of each transposon
+insertion site within each gene.
 
 ```{r run_perl_script}
 system("perl findTA.pl")
@@ -190,7 +193,8 @@ Tables 2, 3, and 4. Table 3 also has an extended version, as
 Supplementary Table 6. There are a number of
 other summary statistics about the data, for example in Table 1 and in
 the right-hand column of the second page of the paper. But I am going
-to focus on the key analysis results, Figures 1-3 and Tables 2-4 and 6.
+to focus on the key analysis results, Figures 1-3 and Tables 2-4 and
+Supplemental Table 6.
 
 Let me first explain some key elements of the project directory in
 [`../original/`](https://github.com/kbroman/Paper_ReScience2020/blob/master/original/). The `R` subdirectory contains my
@@ -238,8 +242,9 @@ save(data100nostop, data100stop, data90, data70, data60, data80,
 setwd("..")
 ```
 
-I will now run the MCMC. (I actually cache these results and only run
-them if needed.) Again, we copied over the
+I will now run the MCMC. This takes about **9 min** on my laptop. (I cache the results and only run
+this if needed.)
+Again, we copied over the
 [`Sept02/R/analysis.R`](https://github.com/kbroman/Paper_ReScience2020/blob/master/original/Sept02/R/analysis.R) script as
 [`R/sept02_analysis.R`](https://github.com/kbroman/Paper_ReScience2020/blob/master/reproduction/R/sept02_analysis.R).
 
@@ -303,7 +308,8 @@ setwd("..")
 ```
 
 And again we run the MCMC. This one is faster, because we are doing
-just one version of the analysis rather than 6. The
+just one version of the analysis rather than 6. (It takes
+about **1.5 min** on my laptop). The
 [`R/nov02_analysis.R`](https://github.com/kbroman/Paper_ReScience2020/blob/master/reproduction/R/nov02_analysis.R) script was copied over from
 [`Nov02/R/analysis.R`](https://github.com/kbroman/Paper_ReScience2020/blob/master/original/Nov02/R/analysis.R).
 
@@ -338,7 +344,7 @@ if(file.exists("R/nov02_results.RData")) {
 
 ## Tables
 
-Let us now reconstruct the tables 2-4 and 6.
+Let us now reconstruct the Tables 2-4 and Supplemental Table 6.
 
 ### Table 2
 
@@ -420,10 +426,11 @@ print(xtab2, file="Tabs/tab2.tex",
 ```
 
 
-### Tables 3 and 6
+### Table 3 and Supplemental Table 6
 
-Let us now turn to Tables 3 and 6, on the genes with probability > 75%
-of being essential.
+Let us now turn to Table 3 and Supplemental Table 6, on the genes with probability > 75%
+of being essential. (Supplemental Table 6 is an extended version of
+Table 3.)
 
 In order to reproduce these tables exactly, we need to make one small
 change to the code in the
@@ -465,7 +472,7 @@ being essential is now 74.8%. This difference is again likely due to
 MCMC sampling error.)
 
 ```{r table3}
-# genes to exclude from Table 3 (just listed in table 6)
+# genes to exclude from Table 3 (just listed in Supplemental Table 6)
 tab3_exclude <- c("418",  "1218", "3003", "1701",
                   "2448", "3002", "1702")
 
