+++ article/metadata.tex
@@ -14,7 +14,7 @@
 \def \dateRECEIVED{}
 \def \dateACCEPTED{}
 \def \datePUBLISHED{}
-\def \articleTITLE{[Rp] Identifying essential genes by mutagenesis}
+\def \articleTITLE{[Rp] Reproducibility report: Identifying essential genes by mutagenesis}
 \def \articleTYPE{Reproduction / Ten Years Reproducibility Challenge}
 \def \articleDOMAIN{Genetics}
 \def \articleBIBLIOGRAPHY{bibliography.bib}
