### 2020-07-21 0cb65b4
Add doi for code and data, plus SWH id
