+++ article/metadata.tex
@@ -1,10 +1,10 @@
 % DO NOT EDIT - automatically generated from metadata.yaml
 
 \def \codeURL{https://github.com/kbroman/Paper_ReScience2020/}
-\def \codeDOI{}
-\def \codeSWH{}
+\def \codeDOI{10.5281/zenodo.3671998}
+\def \codeSWH{swh:1:dir:277cad9553dc627f9a4047f38fc19ca1e76aee8c}
 \def \dataURL{https://github.com/kbroman/Paper_ReScience2020/}
-\def \dataDOI{}
+\def \dataDOI{10.5281/zenodo.3671998}
 \def \editorNAME{Karthik Ram}
 \def \editorORCID{0000-0002-0233-1757}
 \def \reviewerINAME{Leah Wang}
