### 2020-07-24 #18 @karthik
@kbroman Im just waiting to correctly deposit the metadata into Zenodo before I close this out. Stay tuned.
