+++ article/metadata.yaml
@@ -110,5 +110,5 @@ article:
 journal:
   - name:   "ReScience C"
   - issn:   2430-3658
-  - volume:
-  - issue:
+  - volume: 6
+  - issue: 1
