+++ article/metadata.yaml
@@ -102,9 +102,9 @@ dates:
 
 # This information will be provided by the editor
 article:
-  - number: # Article number will be automatically assigned during publication
-  - doi:    # DOI from Zenodo
-  - url:    # Final PDF URL (Zenodo or rescience website?)
+  - number: 12 # Article number will be automatically assigned during publication
+  - doi:   10.5281/zenodo.3959516 # DOI from Zenodo
+  - url:  https://zenodo.org/record/3959516/files/article.pdf  # Final PDF URL (Zenodo or rescience website?)
 
 # This information will be provided by the editor
 journal:
