+++ article/metadata.yaml
@@ -96,9 +96,9 @@ contributors:
 
 # This information will be provided by the editor
 dates:
-  - received:
-  - accepted:
-  - published:
+  - received:  17 Feb 2020
+  - accepted:  15 Jul 2020
+  - published: 24 Jul 2020
 
 # This information will be provided by the editor
 article:
