+++ article/metadata.tex
@@ -11,9 +11,9 @@
 \def \reviewerIORCID{0000-0001-5905-8061}
 \def \reviewerIINAME{Peter Carbonetto}
 \def \reviewerIIORCID{0000-0003-1144-6780}
-\def \dateRECEIVED{}
-\def \dateACCEPTED{}
-\def \datePUBLISHED{}
+\def \dateRECEIVED{17 February 2020}
+\def \dateACCEPTED{15 July 2020}
+\def \datePUBLISHED{24 July 2020}
 \def \articleTITLE{[Rp] Reproducibility report: Identifying essential genes by mutagenesis}
 \def \articleTYPE{Reproduction / Ten Years Reproducibility Challenge}
 \def \articleDOMAIN{Genetics}
@@ -29,10 +29,10 @@
 \def \contactEMAIL{broman@wisc.edu}
 \def \articleKEYWORDS{rescience c, ten years reproducibility challenge, R, MCMC, tuberculosis, Bayes}
 \def \journalNAME{ReScience C}
-\def \journalVOLUME{}
-\def \journalISSUE{}
-\def \articleNUMBER{}
-\def \articleDOI{}
+\def \journalVOLUME{6}
+\def \journalISSUE{1}
+\def \articleNUMBER{12}
+\def \articleDOI{10.5281/zenodo.3959516}
 \def \authorsFULL{Karl W. Broman}
 \def \authorsABBRV{K.W. Broman}
 \def \authorsSHORT{Broman}
