### 2020-07-24 99df2ab
Crap; there was another problem due to R 4.0 stringAsFactors=FALSE
- I knew something was wrong, but I didn't have time to resolve it until today,
  and instead just edited reproduction.html to match the changes to reproduction.Rmd

- **BUT** actually reproduction.Rmd was broken in R version 4.0 due to
  stringsAsFactors=FALSE, as one of the read.csv() calls really assumes that
  one of the string columns will be made a factor.

- So, another ugly change, wherein I paste "`, stringsAsFactors=TRUE`" into the
  data preparation scripts.
