### 2020-08-19 #18 @khinsen
@karthik Since this paper is now published, can we close this issue?
